update graph command to do analysis; remove analyze-graph
