+++ packages/go/stbernard/command/graph/graph.go
@@ -17,6 +17,7 @@ package graph
 
 import (
 	"context"
+	"encoding/json"
 	"flag"
 	"fmt"
 	"io/fs"
@@ -26,12 +27,25 @@ import (
 	"path/filepath"
 	"time"
 
+	"github.com/specterops/bloodhound/cmd/api/src/auth"
+	"github.com/specterops/bloodhound/cmd/api/src/config"
+	"github.com/specterops/bloodhound/cmd/api/src/daemons/datapipe"
+	"github.com/specterops/bloodhound/cmd/api/src/database"
+	"github.com/specterops/bloodhound/cmd/api/src/migrations"
 	"github.com/specterops/bloodhound/cmd/api/src/model"
 	"github.com/specterops/bloodhound/cmd/api/src/services/graphify"
 	"github.com/specterops/bloodhound/cmd/api/src/services/upload"
+	"github.com/specterops/bloodhound/packages/go/graphschema"
+	"github.com/specterops/bloodhound/packages/go/graphschema/ad"
+	"github.com/specterops/bloodhound/packages/go/graphschema/azure"
+	"github.com/specterops/bloodhound/packages/go/graphschema/common"
+	"github.com/specterops/bloodhound/packages/go/lab/generic"
 	"github.com/specterops/bloodhound/packages/go/stbernard/environment"
-	"github.com/specterops/bloodhound/packages/go/stbernard/shared"
+	"github.com/specterops/dawgs"
+	"github.com/specterops/dawgs/drivers/pg"
 	"github.com/specterops/dawgs/graph"
+	"github.com/specterops/dawgs/query"
+	"github.com/specterops/dawgs/util/size"
 )
 
 const (
@@ -41,7 +55,7 @@ const (
 
 type command struct {
 	env     environment.Environment
-	outfile string
+	outpath string
 	path    string
 }
 
@@ -67,7 +81,7 @@ func (s *command) Parse(cmdIndex int) error {
 	var err error
 	cmd := flag.NewFlagSet(Name, flag.ContinueOnError)
 
-	cmd.StringVar(&s.outfile, "outfile", "/tmp/graph.json", "destination path for generic graph file, default is {root}/tmp/graph.json")
+	cmd.StringVar(&s.outpath, "outpath", "/tmp/", "destination path for generic graph files, default is {root}/tmp/")
 	cmd.StringVar(&s.path, "path", "", "directory containing bloodhound collection files")
 
 	cmd.Usage = func() {
@@ -100,25 +114,227 @@ func (s *command) Run() error {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	if graphDB, err := shared.InitializeGraphDatabase(ctx, s.env[environment.PostgresConnectionVarName]); err != nil {
+	if graphDB, err := initializeGraphDatabase(ctx, s.env[environment.PostgresConnectionVarName]); err != nil {
 		return fmt.Errorf("error connecting to graphDB: %w", err)
-	} else if db, err := shared.InitializeDatabase(ctx, s.env[environment.PostgresConnectionVarName]); err != nil {
+	} else if db, err := initializeDatabase(ctx, s.env[environment.PostgresConnectionVarName]); err != nil {
 		return fmt.Errorf("error connecting to database: %w", err)
 	} else if ingestFilePaths, err := s.getIngestFilePaths(); err != nil {
 		return fmt.Errorf("error getting ingest file paths from directory %w", err)
 	} else if err = ingestData(ctx, ingestFilePaths, graphDB); err != nil {
 		return fmt.Errorf("error ingesting data %w", err)
-	} else if nodes, edges, err := shared.GetNodesAndEdges(ctx, graphDB); err != nil {
+	} else if nodes, edges, err := getNodesAndEdges(ctx, graphDB); err != nil {
 		return fmt.Errorf("error retrieving nodes and edges from database %w", err)
-	} else if graph, err := shared.TransformGraph(nodes, edges); err != nil {
+	} else if graph, err := transformGraph(nodes, edges); err != nil {
 		return fmt.Errorf("error transforming nodes and edges to graph %w", err)
-	} else if err := shared.WriteGraphToFile(&graph, s.outfile); err != nil {
+	} else if err := writeGraphToFile(&graph, s.outpath, "ingested.json"); err != nil {
 		return fmt.Errorf("error writing graph to file %w", err)
+	} else if err := datapipe.RunAnalysisOperations(ctx, db, graphDB, config.Configuration{}); err != nil {
+		return fmt.Errorf("error running analysis: %w", err)
+	} else if nodes, edges, err := getNodesAndEdges(ctx, graphDB); err != nil {
+		return fmt.Errorf("error getting nodes and edges: %w", err)
+	} else if graph, err := transformGraph(nodes, edges); err != nil {
+		return fmt.Errorf("error transforming nodes and edges to graph: %w", err)
+	} else if err := writeGraphToFile(&graph, s.outpath, "analyzed.json"); err != nil {
+		return fmt.Errorf("error writing graph to file: %w", err)
 	} else if err := db.Wipe(ctx); err != nil {
 		return fmt.Errorf("error wiping db: %w", err)
 	}
 
 	return nil
+}
+
+func writeGraphToFile(graph *generic.Graph, folder string, fileName string) error {
+	outputFile := filepath.Join(folder, fileName)
+
+	if jsonBytes, err := json.MarshalIndent(generateGenericGraphFile(graph), "", "  "); err != nil {
+		return fmt.Errorf("error occurred while marshalling ingest file into bytes %w", err)
+	} else if err := os.MkdirAll(folder, 0755); err != nil {
+		return fmt.Errorf("creating output directory: %w", err)
+	} else {
+		return os.WriteFile(outputFile, jsonBytes, 0644)
+	}
+}
+
+func generateGenericGraphFile(graph *generic.Graph) generic.GenericObject {
+	return generic.GenericObject{
+		Graph: *graph,
+	}
+}
+
+func transformGraph(nodes []*graph.Node, edges []*graph.Relationship) (generic.Graph, error) {
+	var (
+		isAZBase bool
+		isBase   bool
+
+		graphNodes    = make([]generic.Node, 0, len(nodes))
+		graphEdges    = make([]generic.Edge, 0, len(edges))
+		nodeObjectIDs = make(map[graph.ID]string, len(nodes))
+	)
+
+	for _, node := range nodes {
+		isAZBase = false
+		isBase = false
+
+		var kinds = make([]string, 0, len(node.Kinds))
+
+		for _, kind := range node.Kinds {
+			if kind == ad.Entity {
+				isBase = true
+			} else if kind == azure.Entity {
+				isAZBase = true
+			} else {
+				kinds = append(kinds, kind.String())
+			}
+		}
+
+		if isBase {
+			kinds = append(kinds, ad.Entity.String())
+		} else if isAZBase {
+			kinds = append(kinds, azure.Entity.String())
+		}
+
+		objectID, err := node.Properties.Get(common.ObjectID.String()).String()
+		if err != nil {
+			return generic.Graph{}, err
+		}
+
+		nodeObjectIDs[node.ID] = objectID
+
+		graphNodes = append(graphNodes, generic.Node{
+			ID:         objectID,
+			Kinds:      kinds,
+			Properties: removeNullMapValues(node.Properties.Map),
+		})
+	}
+
+	for _, edge := range edges {
+		graphEdges = append(graphEdges, generic.Edge{
+			Start: generic.Terminal{
+				MatchBy: "id",
+				Value:   nodeObjectIDs[edge.StartID],
+			},
+			End: generic.Terminal{
+				MatchBy: "id",
+				Value:   nodeObjectIDs[edge.EndID],
+			},
+			Kind:       edge.Kind.String(),
+			Properties: removeNullMapValues(edge.Properties.Map),
+		})
+	}
+
+	return generic.Graph{
+		Nodes: graphNodes,
+		Edges: graphEdges,
+	}, nil
+}
+
+func removeNullMapValues[K comparable](m map[K]any) map[K]any {
+	newMap := make(map[K]any, len(m))
+	for k, v := range m {
+		if v == nil {
+			continue
+		}
+
+		newMap[k] = convert(v)
+	}
+	return newMap
+}
+
+func convert(val any) any {
+	switch v := val.(type) {
+	case []any:
+		return removeNullSliceValues(v)
+	case map[any]any:
+		return removeNullMapValues(v)
+	default:
+		return val
+	}
+}
+
+func removeNullSliceValues(l []any) []any {
+	newSlice := make([]any, 0, len(l))
+
+	for _, val := range l {
+		if val == nil {
+			continue
+		}
+
+		newSlice = append(newSlice, convert(val))
+	}
+
+	return newSlice
+}
+
+func getNodesAndEdges(ctx context.Context, database graph.Database) ([]*graph.Node, []*graph.Relationship, error) {
+	var (
+		nodes []*graph.Node
+		edges []*graph.Relationship
+	)
+
+	if err := database.ReadTransaction(ctx, func(tx graph.Transaction) error {
+		err := tx.Nodes().Filter(
+			query.Not(query.Kind(query.Node(), common.MigrationData)),
+		).Fetch(func(cursor graph.Cursor[*graph.Node]) error {
+			for node := range cursor.Chan() {
+				nodes = append(nodes, node)
+			}
+			return cursor.Error()
+		})
+		if err != nil {
+			return fmt.Errorf("error fetching nodes %w", err)
+		}
+		err = tx.Relationships().Fetch(func(cursor graph.Cursor[*graph.Relationship]) error {
+			for edge := range cursor.Chan() {
+				edges = append(edges, edge)
+			}
+			return cursor.Error()
+		})
+		if err != nil {
+			return fmt.Errorf("error fetching relationships %w", err)
+		}
+
+		return nil
+	}); err != nil {
+		return nodes, edges, fmt.Errorf("error occurred reading the database %w", err)
+	} else {
+		return nodes, edges, nil
+	}
+}
+
+func initializeGraphDatabase(ctx context.Context, postgresConnection string) (graph.Database, error) {
+
+	if pool, err := pg.NewPool(postgresConnection); err != nil {
+		return nil, fmt.Errorf("error creating postgres connection %w", err)
+	} else if database, err := dawgs.Open(ctx, pg.DriverName, dawgs.Config{
+		GraphQueryMemoryLimit: size.Gibibyte,
+		ConnectionString:      postgresConnection,
+		Pool:                  pool,
+	}); err != nil {
+		return nil, fmt.Errorf("error connecting to database %w", err)
+	} else if err = migrations.NewGraphMigrator(database).Migrate(ctx, graphschema.DefaultGraphSchema()); err != nil {
+		return nil, fmt.Errorf("error migrating graph %w", err)
+	} else if err = database.SetDefaultGraph(ctx, graphschema.DefaultGraph()); err != nil {
+		return nil, fmt.Errorf("error setting default graph %w", err)
+	} else {
+		return database, nil
+	}
+
+}
+
+func initializeDatabase(ctx context.Context, connection string) (database.Database, error) {
+	var db database.Database
+
+	if gormDB, err := database.OpenDatabase(connection); err != nil {
+		return db, fmt.Errorf("error opening database %w", err)
+	} else {
+		db = database.NewBloodhoundDB(gormDB, auth.NewIdentityResolver())
+	}
+
+	if err := db.Migrate(ctx); err != nil {
+		return db, fmt.Errorf("error migrating database %w", err)
+	}
+
+	return db, nil
 }
 
 func ingestData(ctx context.Context, filepaths []string, database graph.Database) error {
